[FP-76] Refactor: 체험 상세 페이지: 엑세스 토큰 만료시 에러 메세지 수정
