+++ src/constants/errorMessages.ts
@@ -19,7 +19,7 @@ export default ErrorMessages;
 export const DELETE_ACTIVITY_ERROR_MESSAGES = {
   PENDING_RESERVATION: '신청 예약이 있는 체험은 삭제할 수 없습니다.',
   CONFIRMED_RESERVATION: '확정 예약이 있는 체험은 삭제할 수 없습니다.',
-  UNAUTHORIZED: '인증되지 않은 사용자입니다.',
+  UNAUTHORIZED: '로그인 만료 시간이 지났습니다.',
   FORBIDDEN: '본인의 체험만 삭제할 수 있습니다.',
   NOT_FOUND: '존재하지 않는 체험입니다.',
   DEFAULT: '체험을 삭제하는데 실패했습니다.',
